Add a function "defaultLayout'" that works as same as defaultLayout in Yesod type class.
"defautlLayout'" is used instead of "defaultLayout" to prevent from infinit. recursion.
Type Yesod defines "defaultLayout", and it allows ad-hook overloding of it.
However, "defaultLayout" is not ad-hook overloaded in the definition
in the instance of Yesod type class.
"defaultLayout'", which is different definition with "defaultLayout", can avoid infinite
recursion.
That's the reason why "defaultLayout'" is necessary.